Update to be almost stable
This patch adds a few things to the Dockerfile to push it further
towards a stable state. Below are the few things left to do before
this repo can be called "stable".
TODO: Remove `export O...` from contents/configure
TODO: Add ssh-keygen ... to Makefile
TODO: Find a way to log root in when container starts